// src/types/build-log.ts
export interface buildLogTypes {
  /**
   * 作者列表
   */
  authorList?: string[]
  /**
   * 分支列表
   */
  branchList?: string[]
  /**
   * node 环境
   */
  environment?: string
  /**
   * node 环境后缀
   */
  environmentSuffix?: string
  /**
   * 多应用项目名称
   */
  project?: string
}